requirements.txt
aiohttp==3.10.2
python-dotenv==1.0.1
